+++ Embedded_Code/DAS/data.csv
@@ -1,0 +1,5 @@
+-0.044896, -0.093025, 1.024007, 1.032500, 0.516250, -5.302500, 1023.000000, 1023.000000, 137.913406, 0.000000
+-0.042151, -0.107665, 1.031632, 1.128750, 0.673750, -5.232500, 1023.000000, 1023.000000, 44.639027, 0.000000
+-0.031842, -0.081496, 1.008818, 1.697500, 49.376251, -13.011250, 1023.000000, 1023.000000, 153.319077, 0.000000
+-0.044042, -0.088877, 1.015467, 1.023750, 0.700000, -5.328750, 1023.000000, 1023.000000, 42.398201, 0.000000
+-0.046482, -0.092842, 1.025593, 0.962500, 0.743750, -5.337500, 1023.000000, 1023.000000, 35.955826, 0.000000
