+++ Embedded_Code/SSA/finalssa/main/main.cpp
@@ -20,19 +20,22 @@ void setup() //initializes different sensors
   if(!initSD()) { 
     Serial.print("SD initialization failed");  
   }
+  SdRemove(); //removes existing file
 }
 
 void loop() //Eventually going to want to multithread this so the other threads can make progress while wheel speed delays
 {	
   analogData(&temparr[0]);
   lsm9ds1.getAccelData(); //Read accel data
-  //lsm9ds1.getGyroData(); //Read gyro data
-  //lsm9ds1.printAccelData();
-//	wheelspeedSetup();
-//  while(getwheelspeedData() == 0) { continue; } //waiting for magnet to trigger, magnet has to trigger in order for execution to finish
-//  wheelspeed = getwheelspeedData();
-  printAllData();
-  SdWrite(lsm9ds1);
+  lsm9ds1.getGyroData(); //Read gyro data
+	wheelspeedSetup();
+  while(getwheelspeedData() == 0) { continue; } //waiting for magnet to trigger, magnet has to trigger in order for execution to finish
+  wheelspeed = getwheelspeedData();
+//  printAllData();
+  
+  if(!SdWrite(lsm9ds1, temparr[0], temparr[1], temparr[2], 0.0)) { 
+    Serial.print("Couldn't open sd card");  
+  }
   delay(100);
 }
 
@@ -47,7 +50,6 @@ void printAllData()
   Serial.println(temparr[2]);
   Serial.println("\n");
 
-  /*
   //Accelerometer data
   Serial.print("X axis accel: ");
   Serial.println(lsm9ds1.getAccelX());
@@ -71,5 +73,5 @@ void printAllData()
   Serial.print("WheelSpeed: ");
   Serial.println(wheelspeed);
   Serial.println("\n");
-  */
+  
 }
