+++ Embedded_Code/SSA/finalssa/sd/sd.cpp
@@ -20,13 +20,26 @@ bool initSD()
    return true;
 }
 
-void SdWrite(IMU imu)
+void SdRemove()
+{
+   SD.remove("data.csv");
+}
+
+bool SdWrite(IMU imu, float t1, float t2, float t3, float ws)
 {
    //check and remove existing file
    file = SD.open("data.csv", FILE_WRITE);
-   file.print("Hello John\n");
-   SdClose();
-   //write data formatted to csv
+   if(!file) { return false; }
+   else {
+      char fstr[50]; //conservative buffer size
+      sprintf(fstr, "%f, %f, %f, %f, %f, %f, %f, %f, %f, %f\n", imu.getAccelX(), imu.getAccelY(), imu.getAccelZ(),\
+      imu.getGyroX(), imu.getGyroY(), imu.getGyroZ(), t1, t2, t3, ws); //formatting string
+      file.seek(EOF); //gets to end of file so append occurs
+      file.println(fstr); //printing to string
+   }
+   SdClose(); //opening and closing file each time to minimize risk of corruption due to unfortunate shutdown
+   //TODO: Receive CAN signal from CCM before shutdown so file can be closed / opened only once
+   return true;
 }
 
 void SdClose()
